+++ CudaBeamformer/Microphone/Microphone.h
@@ -8,30 +8,30 @@ using leapIter = leapMap::iterator;
 class Microphone
 {
 public:
-	Microphone( double distCenter ) ;
+	Microphone( float distCenter ) ;
 	~Microphone();
 
-	double getSteeringDelay(double steeringAngle) const
+	float getSteeringDelay(float steeringAngle) const
 	{
-		double returnVal = -m_distCenter * sin(steeringAngle * GLOBAL_PI / 180.0) / GLOBAL_SOUND_SPEED * (double)Config::getInstance().samplePerSecond;
+		float returnVal = -m_distCenter * sin(steeringAngle * GLOBAL_PI / 180.0) / GLOBAL_SOUND_SPEED * (float)Config::getInstance().samplePerSecond;
 		return returnVal;
 	}
 
-	double getFocusDelay(double focusDist) const
+	float getFocusDelay(float focusDist) const
 	{
 		if (focusDist < 0.1)
 			return 0;
-		double returnVal = (pow(m_distCenter, 2) / (2.0 * focusDist)) / GLOBAL_SOUND_SPEED * (double)Config::getInstance().samplePerSecond;
+		float returnVal = (pow(m_distCenter, 2) / (2.0 * focusDist)) / GLOBAL_SOUND_SPEED * (float)Config::getInstance().samplePerSecond;
 		return returnVal;
 	}
 
-	double getDelay(double focusDist, double steeringAngle) const
+	float getDelay(float focusDist, float steeringAngle) const
 	{
 		return getFocusDelay(focusDist) + getSteeringDelay(steeringAngle);
 	}
 
 
-	void feed(const SharpVector& input, double focusDist, double steeringAngle, int speakerID)
+	void feed(const SharpVector& input, float focusDist, float steeringAngle, int speakerID)
 	{
 		size_t delay = getDelay(focusDist, steeringAngle) + Config::getInstance().getMicMaxDelay();
 		auto beginIter = input.data();
@@ -45,19 +45,19 @@ public:
 			}
 			else if (k < m_data.size())
 			{
-				double soundData = *beginIter++;
+				float soundData = *beginIter++;
 				m_data[k] += soundData;
 			}
 			else
 			{
-				double soundData = *beginIter++;
+				float soundData = *beginIter++;
 				leapIte->second->at(k - m_data.size()) = soundData;
 				m_leapTotalData[k - m_data.size()] += soundData;
 			}
 		}
 	}
 
-	leapIter getLeapIter(int speakerID, double delay)
+	leapIter getLeapIter(int speakerID, float delay)
 	{
 		leapIter leapIte = m_leapData.find(speakerID);
 		if (leapIte == m_leapData.end())
@@ -90,7 +90,7 @@ public:
 		return m_leapTotalData;
 	}
 
-	double getData(size_t index) const
+	float getData(size_t index) const
 	{
 		if (m_data.size() > index)
 		{
@@ -105,7 +105,7 @@ public:
 
 private:
 
-	double m_distCenter;
+	float m_distCenter;
 	std::unordered_map<int, std::shared_ptr<SharpVector>>  m_leapData; // Leap Data for each source
 	SharpVector m_leapTotalData; // Leap Data for each source
 	SharpVector m_data;
