+++ CudaBeamformer/Microphone/MicrophoneArray.cpp
@@ -4,16 +4,16 @@
 MicrophoneArray::MicrophoneArray()
 {
 	int arraySize = Config::getInstance().arraySize;
-	double elemDistFromMid = arraySize/2;
+	float elemDistFromMid = arraySize/2;
 	for (int i = 0; i < arraySize; i++)
 	{
 		if (arraySize % 2 == 0)
 		{
-			micropshoneList.emplace_back((double(i - elemDistFromMid) + 0.5) * Config::getInstance().distBetweenMics);
+			micropshoneList.emplace_back((float(i - elemDistFromMid) + 0.5) * Config::getInstance().distBetweenMics);
 		}
 		else
 		{
-			micropshoneList.emplace_back((double(i - elemDistFromMid)) * Config::getInstance().distBetweenMics);
+			micropshoneList.emplace_back((float(i - elemDistFromMid)) * Config::getInstance().distBetweenMics);
 		}
 	}
 }
@@ -22,7 +22,7 @@ MicrophoneArray::~MicrophoneArray()
 {
 }
 
-void MicrophoneArray::InsertSound( const std::vector<double>& rawData, double focusDist, double steerAngle)
+void MicrophoneArray::InsertSound( const std::vector<float>& rawData, float focusDist, float steerAngle)
 {
 	for (auto& microphone : micropshoneList)
 	{
@@ -30,7 +30,7 @@ void MicrophoneArray::InsertSound( const std::vector<double>& rawData, double fo
 	}
 }
 
-void MicrophoneArray::Beamform(std::vector<double>& outputData, double focusDist, double steerAngle)
+void MicrophoneArray::Beamform(std::vector<float>& outputData, float focusDist, float steerAngle)
 {
 	std::fill(outputData.begin(), outputData.end(), 0.0);
 	for (size_t i = 0; i < Config::getInstance().packetSize; i++)
